+++ BudgetBuddies/BudgetBuddies/Sources/Network/Calendar/Router/DiscountInfoRouter.swift
@@ -1,0 +1,17 @@
+//
+//  DiscountInfoRouter.swift
+//  BudgetBuddies
+//
+//  Created by 김승원 on 8/9/24.
+//
+
+import Foundation
+import Moya
+
+enum DiscountInfoRouter {
+    case getDiscounts(request: InfoRequest)
+}
+
+//extension DiscountInfoRouter: TargetType {
+//    
+//}
