+++ BudgetBuddies/BudgetBuddies/Sources/Network/Calendar/Router/InfoRequest.swift
@@ -1,0 +1,15 @@
+//
+//  InfoRequest.swift
+//  BudgetBuddies
+//
+//  Created by 김승원 on 8/9/24.
+//
+
+import Foundation
+
+struct InfoRequest {
+    let year: Int
+    let month: Int
+    let page: Int
+    let size: Int
+}
